Fix load/undo/redo
The problem with undo/redo was, that a shallow clone was performed,
so the gems on the tiles were not reset.

The problem with load was that the player references did not point to
the same player, even though they had the same name.

By using deep clones and fixing the player references afterwards,
undo/redo should be fixed (hopefully).

For loading the game from a file, it is enough to fix the player
references once loaded, as loading from JSON behaves like a deep clone.

Another issue I fixed is that the redo stack was not cleared
when a player placed a tile, as it is possible that
we diverged from the redo stack with that placed tile.