Added placeTileAi
placeTileAi is called either at the end of startGame() if the first playerAtTurn is Ai, or it is called at the end of changePlayer(), where it checks if the new PlayerAtTurn is Ai.

PlaceTileAi will play the move after aiMoveMilliseconds

Issue: Although the Tiles are positioned on the board, they remain invisible in the GUI. To verify their placement, you can click on a Tile to view the available positions, which will then reveal the invisibly placed Tiles.